+++ lib/job-iteration/active_record_enumerator.rb
@@ -7,10 +7,20 @@ module JobIteration
   class ActiveRecordEnumerator
     SQL_DATETIME_WITH_NSEC = "%Y-%m-%d %H:%M:%S.%N"
 
-    def initialize(relation, columns: nil, batch_size: 100, timezone: nil, cursor: nil, instance: nil, instances: nil)
+    def initialize(
+      relation,
+      columns: nil,
+      batch_size: 100,
+      timezone: nil,
+      cursor: nil,
+      instance: nil,
+      instances: nil,
+      around_query: nil
+    )
       @relation = relation
       @batch_size = batch_size
       @timezone = timezone
+      @around_query = around_query
       @columns = if columns
         Array(columns)
       else
@@ -64,7 +74,14 @@ module JobIteration
     end
 
     def finder_cursor
-      JobIteration::ActiveRecordCursor.new(@relation, @columns, @cursor, @instance, @instances)
+      JobIteration::ActiveRecordCursor.new(
+        @relation,
+        @columns,
+        @cursor,
+        @instance,
+        @instances,
+        around_query: @around_query,
+      )
     end
 
     def column_value(record, attribute)